// src/interfaces/companyUser.ts
export interface CompanyUser {
    image: string;
    companyName: string;
    role: string;
    phoneNumber: string;
    address: string;
    postalCode: string;
    state: string;
    city: string;
    description: string;
    youtubeLink: string;
    images: string[];
    promotions: string;
    discounts: string;
    googleMapsLink: string;
    email: string;
    password: string;
}